# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/clubstats": {
            "get": {
                "description": "Get statistics for a club based on name and index",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get club statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Club Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ClubStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/playerstats": {
            "get": {
                "description": "Get statistics for a player based on name, index, and nationality",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get player statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Player Name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Index",
                        "name": "index",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Nationality",
                        "name": "nationality",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PlayerStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "model.ClubStats": {
            "type": "object",
            "properties": {
                "last_games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.LastGames"
                    }
                },
                "market_value": {
                    "type": "string"
                },
                "next_games": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NextGames"
                    }
                },
                "result_stats": {
                    "$ref": "#/definitions/model.ResultStats"
                },
                "result_stats_per_competition": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.ResultStats"
                    }
                },
                "season": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "model.LastGames": {
            "type": "object",
            "properties": {
                "away_team": {
                    "type": "string"
                },
                "competition": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "form": {
                    "type": "string"
                },
                "home_team": {
                    "type": "string"
                },
                "hour": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "model.NextGames": {
            "type": "object",
            "properties": {
                "away_team": {
                    "type": "string"
                },
                "competition": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "home_team": {
                    "type": "string"
                },
                "hour": {
                    "type": "string"
                }
            }
        },
        "model.PlayerClubStats": {
            "type": "object",
            "properties": {
                "assists": {
                    "type": "integer"
                },
                "club": {
                    "type": "string"
                },
                "goals_scored": {
                    "type": "integer"
                },
                "matches_played": {
                    "type": "integer"
                }
            }
        },
        "model.PlayerStats": {
            "type": "object",
            "properties": {
                "player_name": {
                    "type": "string"
                },
                "seasons": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/model.PlayerClubStats"
                        }
                    }
                },
                "tournaments": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.TournamentStats"
                    }
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ResultStats": {
            "type": "object",
            "properties": {
                "defeats": {
                    "type": "integer"
                },
                "draws": {
                    "type": "integer"
                },
                "goals_conceded": {
                    "type": "integer"
                },
                "goals_scored": {
                    "type": "integer"
                },
                "matches_played": {
                    "type": "integer"
                },
                "victories": {
                    "type": "integer"
                }
            }
        },
        "model.TournamentStats": {
            "type": "object",
            "properties": {
                "assists": {
                    "type": "integer"
                },
                "goals_scored": {
                    "type": "integer"
                },
                "matches_played": {
                    "type": "integer"
                }
            }
        }
    }
}